# ELP_App/elp-app-node/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ELP App API",
    "description": "API documentation for the ELP App"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
...
    "/user/session": {
      "post": {
        "summary": "Authenticate user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userid": { "type": "string" },
                "password": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "token": { "type": "string" },
                "userid": { "type": "string" },
                "role": { "type": "string" },
                "name": { "type": "string" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/user/session/current": {
      "delete": {
        "summary": "Logout user",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Authentication token"
          }
        ],
        "responses": {
          "200": { "description": "Logout successful" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}